NFS: fix dynamic status file length getting out of sync with reality
NFSv4 has a nicer model here with actual opens it seems, but with
NFSv3 there doesn't seem to be a great way to do dynamic files.
Instead, just regenerate it every few seconds and cache it hope for
the best.

In practice, this means the GetAttr/LStat before the Read will generate it and
then the bytes will be cached for the read and the count will match.

Previously the total length was often wrong by a 1 byte short or a
byte long (due to nanoseconds in proc uptime field) and then the JSON
wouldn't be valid.

We can't just set size 0 because then the NFS client doesn't even
issue reads.  So we seem to need to know the real length ahead of
time.

Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>