testing: flesh out winfsp, add to CI, expand CI, fix gitstore bug
This runs "go mod verify" against gomodfs in NFS mode on all three
OSes (linux, darwin, windows) and additionally runs Windows in WinFsp
mode.

In the process, this also implements the magic ".gomodfs-status" file
work in PowerShell with WinFsp. (Powershell sets FILE_OPEN_NO_RECALL
which go-winfsp was rejecting.  So this adds a wrapper that unsets
that bit.)

And this fixes a gitstore bug that made "go mod verify" fail on a repo
with exotic filenames. (#15)

Updates tailscale/corp#34812
Fixes #15

Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>